requirements.txt
langchain
fastapi
uvicorn
jinja2
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-openai
langchain-community
